+++ crowd-fund-analysis/cf_analysis_agent/api/public_equity_api.py
@@ -1,6 +1,6 @@
 import json
 from flask import Blueprint, request, jsonify
-from cf_analysis_agent.utils.get_criteria import generate_ai_criteria ,upload_ai_criteria_to_s3, update_custom_criteria,fetch_criteria_file
+from cf_analysis_agent.utils.criteria_utils import generate_ai_criteria ,upload_ai_criteria_to_s3, update_custom_criteria,fetch_criteria_file
 from cf_analysis_agent.structures.criteria_structures import StructuredIndustryGroupCriteriaResponse, EquityDeatils, IndustryGroupData
