+++ crowd-fund-analysis/cf_analysis_agent/utils/criteria_utils.py
